# docs/openapi(2) (1).yaml
openapi: 3.0.3
info:
  title: Onidel Cloud API - OpenAPI 3.0
  description: |-
    Welcome to the **Onidel Cloud API** documentation. This API allows developers to seamlessly integrate with the Onidel Cloud platform, enabling them to manage cloud resources efficiently and securely. With our API, you can automate processes, retrieve detailed account information, and manage virtual machines (VPS), networking, storage, and more.
  termsOfService: https://docs.onidel.com/policies/terms-of-service
  contact:
    email: support@onidel.com
  version: 1.2.0
servers:
  - url: https://api.cloud.onidel.com
...
paths:
...
  /snapshots/{snapshot_id}:
    post:
      tags:
        - Snapshots
      summary: Generate snapshot download URL
      description: Generate snapshot download URL
      operationId: genSnapshotURL
      security:
        - api_key: []
      parameters:
        - name: snapshot_id
          in: path
          description: Snapshot UUID that need to be considered for filter
          required: true
          explode: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotDownloadURL'
        '401':
          description: Unauthorized
        '400':
          description: Bad request
    delete:
      tags:
        - Snapshots
      summary: Delete a snapshot
      description: Delete a snapshot
      operationId: deleteSnapshot
      security:
        - api_key: []
      parameters:
        - name: snapshot_id
          in: path
          description: Snapshot UUID that need to be considered for filter
          required: true
          explode: true
          schema:
            type: string
            format: uuid
        - name: team_id
          in: query
          description: Team ID that need to be considered for filter. Default team will be used if not provided.
          required: false
          explode: true
          schema:
            type: string
      responses:
        '204':
          description: Successful operation
        '401':
          description: Unauthorized
        '404':
          description: Not found
...
components:
  schemas:
...
    SnapshotDownloadURL:
      type: object
      properties:
        url:
          type: string
          format: URL
          example: https://syd.dl.onidel.com/download?fid=...
...
  securitySchemes:
    api_key:
      type: http
      scheme: bearer